fix types

+++ ProjectRelativity/Item.cs
@@ -1,3 +1,3 @@
 namespace ProjectRelativity;
 
-public record Item(int id, string name, int price);
+public record Item(int id, string name, double price);
